Add simple add-to-list functionality for game details also simplified list
Add simple add-to-list functionality for game details also simplified list. So that i can first get the list working with basic information. In gamedetails i
Properly made the game status added values to it to easier add type to const.
Handle "Add to List" button click: send selected game, status, and score to backend via POST
Show alert on success or if user is not logged in